+++ tests/test_cats.py
@@ -1,10 +1,16 @@
 import typing as ty
 from datetime import datetime
 
-import pytest
 import attr
-
-from typecats import Cat, unstruc, struc, register_struc_hook, register_unstruc_hook
+import pytest
+from typecats import (
+    Cat,
+    StructuringError,
+    register_struc_hook,
+    register_unstruc_hook,
+    struc,
+    unstruc,
+)
 
 
 @Cat
@@ -17,13 +23,13 @@ class CatTest:
 
 def test_cats_decorator() -> None:
 
-    with pytest.raises(TypeError):
+    with pytest.raises(StructuringError):
         # missing neutered
         CatTest.struc(dict(name="Tom", age=2, alive=False))
-    with pytest.raises(TypeError):
+    with pytest.raises(StructuringError):
         # missing age
         CatTest.struc(dict(name="Tom", neutered=False, alive=False))
-    with pytest.raises(ValueError):
+    with pytest.raises(StructuringError):
         # missing non-empty name
         CatTest.struc(dict(name="", age=2, neutered=True, alive=True))
 
@@ -59,15 +65,11 @@ class Task:
 
 def test_nested_with_structurer() -> None:
 
-    register_struc_hook(
-        datetime, lambda s, _t: datetime.strptime(s, "%Y-%m-%dT%H:%M:%S.%fZ")
-    )
+    register_struc_hook(datetime, lambda s, _t: datetime.strptime(s, "%Y-%m-%dT%H:%M:%S.%fZ"))
     register_unstruc_hook(datetime, lambda d: d.isoformat() + "Z")
 
     dts = "2019-06-27T13:04:04.000111Z"
-    dict_task = dict(
-        name="feed cats", created_at=dts, subtasks=[dict(name="empty bowl")]
-    )
+    dict_task = dict(name="feed cats", created_at=dts, subtasks=[dict(name="empty bowl")])
     task = Task.struc(dict_task)
 
     assert task.subtasks[0] == Subtask("empty bowl")