+++ tests/test_exceptions.py
@@ -1,12 +1,32 @@
 from typing import List
 
 import pytest
-
 from typecats import Cat
-from typecats.patch import CommonStructuringExceptions
+from typecats.exceptions import StructuringError
+
+from detailed_validation_utils import unsafe_stack_disable_detailed_validation
 
 
-def test_exceptions_get_logged(caplog):
+def test_exceptions_dont_have_type_path(caplog):
+    """All offending type paths are reported with detailed validation by cattrs.
+    This makes this feature redundant, so we don't bother including the offending paths.
+    """
+
+    with pytest.raises(StructuringError):
+        Zap.struc(
+            dict(
+                x=3,
+                y=55,
+                foo_matrix=[[dict(baz="a"), dict(baz="b")], [dict(baz="c")], [4]],
+            )
+        )
+
+    rec = caplog.records.pop(0)
+    assert "Failed to structure Zap from item <{'x': 3," in rec.msg
+    assert "at type path" not in rec.msg
+
+
+def test_exceptions_get_logged_no_detailed_validation(caplog):
     @Cat
     class Bar:
         baz: int
@@ -16,8 +36,9 @@ def test_exceptions_get_logged(caplog):
         foo: str
         bar: Bar
 
-    with pytest.raises(CommonStructuringExceptions):
-        Quux.struc(dict(foo="foos", bar=[1, 2, 3]))
+    with pytest.raises(StructuringError):
+        with unsafe_stack_disable_detailed_validation():
+            Quux.struc(dict(foo="foos", bar=[1, 2, 3]))
 
     rec = caplog.records.pop(0)
     assert rec.msg.startswith("Failed to structure Bar ")
@@ -36,19 +57,19 @@ class Zap:
     foo_matrix: List[List[Foo]]
 
 
-def test_exceptions_have_type_path(caplog):
+def test_exceptions_have_type_path_no_detailed_validation(caplog):
 
-    with pytest.raises(CommonStructuringExceptions):
-        Zap.struc(
-            dict(
-                x=3,
-                y=55,
-                foo_matrix=[[dict(baz="a"), dict(baz="b")], [dict(baz="c")], [4]],
+    with pytest.raises(StructuringError):
+        with unsafe_stack_disable_detailed_validation():
+            Zap.struc(
+                dict(
+                    x=3,
+                    y=55,
+                    foo_matrix=[[dict(baz="a"), dict(baz="b")], [dict(baz="c")], [4]],
+                )
             )
-        )
 
     rec = caplog.records.pop(0)
-    print(rec.msg)
     assert "Failed to structure Foo from item <4> at type path" in rec.msg
 
 
@@ -57,3 +78,16 @@ def test_try_struc_no_exception_if_common(caplog):
     assert None is Zap.try_struc(dict(x=4, y=55, foo_matrix=[[[]]]))
 
     assert not caplog.records
+
+
+def test_cache_is_cleared_when_changing_validation_mode(caplog):
+    """If something is structured _once_ with detailed validation enabled, disabling it has no effect
+    on the existing cached structuring functions, as their generated code will continue to "collect"
+    detailed validation errors which are grouped and won't raise errors where they actually occurred in the stack,
+    essentially bypassing _embed_exception_info."""
+
+    test_exceptions_dont_have_type_path(caplog)
+    test_exceptions_have_type_path_no_detailed_validation(caplog)
+    test_exceptions_dont_have_type_path(caplog)
+    test_exceptions_have_type_path_no_detailed_validation(caplog)
+    test_exceptions_dont_have_type_path(caplog)