+++ typecats/__about__.py
@@ -1,4 +1,4 @@
 """typecats"""
-__version__ = "1.7.1"
+__version__ = "2.0.0rc1"
 __author__ = "Peter Gaultney"
 __author_email__ = "pgaultney@xoi.io"