+++ typecats/__init__.py
@@ -5,10 +5,16 @@ from .tc import (  # noqa
     try_struc,
     register_struc_hook,
     register_unstruc_hook,
+    register_struc_hook_func,
+    register_unstruc_hook_func,
+    unstruc_strip_defaults,
+    set_detailed_validation_mode_not_threadsafe,
 )
+
+from .exceptions import StructuringError  # noqa
+
 from .wildcat import is_wildcat  # noqa
 from .patch import patch_converter_for_typecats  # noqa
-from .strip_defaults import unstruc_strip_defaults  # noqa
 
 from .__about__ import __version__