+++ typecats/strip_defaults.py
@@ -1,13 +1,14 @@
 import typing as ty
 from functools import lru_cache
-from typing_extensions import Literal
+import contextvars as cv
 
 import attr
-from cattr.converters import _is_attrs_class, Converter
+from attr import has as is_attrs_class
 
-from .patch import TypecatsCattrPatch
+from ._compat import Literal
 
-C = ty.TypeVar("C")
+
+ShouldStripDefaults = cv.ContextVar("TypecatsShouldStripDefaults", default=False)
 
 
 _MISSING = object()
@@ -24,27 +25,6 @@ def _get_attr_default_value(attribute) -> ty.Any:
     return _get_factory_default(attribute)
 
 
-def _strip_attr_defaults(
-    attrs_type: ty.Type, m: ty.Mapping[str, ty.Any]
-) -> ty.Dict[str, ty.Any]:
-    """The idea here is that when you are using pure dicts, a key can be
-    missing to indicate absence.  But if you're dealing with typed
-    objects, that's not possible since all keys are always present.  So
-    the only 'reasonable' way to determine what a 'union' means in a
-    class-based world is to prefer non-default values to default values at
-    all times, which attrs can tell us about.
-    """
-    attr_defaults_by_name = {
-        attribute.name: _get_attr_default_value(attribute)
-        for attribute in attrs_type.__attrs_attrs__
-    }
-    return {
-        k: v
-        for k, v in m.items()
-        if k not in attr_defaults_by_name or v != attr_defaults_by_name[k]
-    }
-
-
 def _get_names_of_defaulted_nonliteral_attrs(attrs_obj: ty.Any) -> ty.Set[str]:
     res: ty.Set[str] = set()
     for _attr in attrs_obj.__attrs_attrs__:
@@ -56,33 +36,15 @@ def _get_names_of_defaulted_nonliteral_attrs(attrs_obj: ty.Any) -> ty.Set[str]:
     return res
 
 
-def _strip_attrs_defaults(
-    unstructured_but_unclean: ty.Any, obj_to_unstructure: ty.Any
-) -> ty.Any:
-    if _is_attrs_class(obj_to_unstructure.__class__):
-        keys_to_strip = _get_names_of_defaulted_nonliteral_attrs(obj_to_unstructure)
-        return {
-            k: v for k, v in unstructured_but_unclean.items() if k not in keys_to_strip
-        }
-    return unstructured_but_unclean
+def strip_attrs_defaults(unstructured_but_unclean: ty.Any, obj_to_unstructure: ty.Any) -> ty.Any:
+    """The idea here is that when you are using pure dicts, a key can be
+    missing to indicate absence.  But if you're dealing with typed
+    objects, that's not possible since all keys are always present.  So
+    the only 'reasonable' way to determine what a 'union' means in a
+    class-based world is to prefer non-default values to default values at
+    all times, which attrs can tell us about.
+    """
 
-
-class StripAttrsDefaultsOnUnstructurePatch(TypecatsCattrPatch):
-    def unstructure_patch(
-        self, original_handler: ty.Callable, obj_to_unstructure: ty.Any
-    ) -> ty.Any:
-        rv = super().unstructure_patch(original_handler, obj_to_unstructure)
-        return _strip_attrs_defaults(rv, obj_to_unstructure)
-
-
-_STRIP_DEFAULTS_CONVERTER = Converter()  # create converter
-__PATCH = StripAttrsDefaultsOnUnstructurePatch(_STRIP_DEFAULTS_CONVERTER)  # patch it
-
-
-def get_stripping_converter() -> Converter:
-    return _STRIP_DEFAULTS_CONVERTER
-
-
-def unstruc_strip_defaults(obj: ty.Any) -> ty.Any:
-    """This is the only thing anyone outside needs to worry about"""
-    return _STRIP_DEFAULTS_CONVERTER.unstructure(obj)
+    assert is_attrs_class(obj_to_unstructure.__class__)
+    keys_to_strip = _get_names_of_defaulted_nonliteral_attrs(obj_to_unstructure)
+    return {k: v for k, v in unstructured_but_unclean.items() if k not in keys_to_strip}