+++ CHANGES.md
@@ -1,8 +1,25 @@
-### 1.7.1
+# 2.0.0
+
+Breaking changes:
+
+- Upgrades `cattrs` from 1.1.2 to 22.1.0, `attrs` from 20.3.0 to 21.4.0 and switches to the
+  GenConverter, which supports the newer style type annotations (e.g, list[] instead of typing.List[]).
+- Detailed validation from cattrs is enabled by default. Disable it if you wish by calling `set_detailed_validation_mode_not_threadsafe(enable=False)`
+- Removed `typecats.types.CommonStructuringExceptions`. Structuring validation errors are now all `typecats.StructuringError`s (aliased to `cattrs.errors.BaseValidationError`), regardless of detailed validation.
+- MyPy plugin changed from resolving from `__builtins__.<type>` to `builtins.<type>` to avoid errors. This might not work on older mypy versions.
+
+Other changes:
+
+- Exports `register_struc_hook_func` and `register_unstruc_hook_func`, which are methods bound to typecats's default cattrs converter.
+- Changed imports from `cattr` to `cattrs`
+- Supports Python 3.8 and 3.9.
+- Changed function parameter types from `cattrs.Converter` to `cattrs.GenConverter` where necessary, though typecats assumes `GenConverter` throughout, so you should probably update too.
+
+## 1.7.1
 
 - Correctly reference typing_extensions as a dependency in setup.py
 
-### 1.7.0
+## 1.7.0
 
 - Upgrades `cattrs` from 1.0.0 to 1.1.2 and `attrs` from 19.1.0 to 20.3.0,
   primarily to fix a `cattrs` bug wherein union field types where types in the
@@ -82,7 +99,7 @@ The non-empty validator will still be run on attributes without defaults.
 ## 1.1.0
 
 No longer block `__setitem__`, `__getitem__`, and `update` for keys
-where the Wildcat class defines an attribute with the same name.  Log
+where the Wildcat class defines an attribute with the same name. Log
 warnings instead. This more permissive behavior will allow a smoother
 progression for code that was written before an attribute was
 subsequently added (typed).